+++ packages/apostrophe-astro/helpers/server/fetch.js
@@ -1,7 +1,5 @@
 import config from 'apostrophe-astro-config/config';
 import { getAposHost } from './url.js';
-import aposResponse from '../../lib/aposResponse.js';
-import aposRequest, { isAstroPrerenderedRequest } from '../../lib/aposRequest.js';
 
 /**
  * A transparent proxy around the native `fetch` API for **server-side
@@ -53,101 +51,3 @@ export async function aposFetch(input, init) {
   });
 }
 
-/**
- * Fetch a full Apostrophe page data object for the given Astro request.
- *
- * This is the primary entry point for SSR and static-build page routes.
- * It wraps `aposRequest` and `aposResponse` to forward the incoming
- * request to the Apostrophe backend and return the parsed JSON page data,
- * including automatic handling of trailing-slash redirects.
- *
- * For static builds, use this inside `getStaticPaths` / your page
- * frontmatter to retrieve the `aposData` prop.
- *
- * @param {Request} req - The incoming Astro request (`Astro.request`).
- * @returns {Promise<object>} The Apostrophe page data object. On error,
- *   returns an object with `errorFetchingPage` set to the caught error
- *   and `page.type` set to `'apos-fetch-error'`.
- *
- * @example
- * ```astro
- * ---
- * import { aposPageFetch } from '@apostrophecms/apostrophe-astro/helpers/server';
- * const aposData = await aposPageFetch(Astro.request);
- * ---
- * ```
- */
-export async function aposPageFetch(req) {
-  let aposData = {};
-  try {
-    // Pass only the URL (as a plain string) when the request is Astro's
-    // prerendered request. Astro v6 installs a warning getter on
-    // Astro.request.headers for prerendered pages; we detect this via
-    // isAstroPrerenderedRequest() which inspects the property descriptor
-    // without triggering the getter. Fall back to env-var / config checks
-    // for any edge cases where the request was already unwrapped.
-    const isStaticBuild = isAstroPrerenderedRequest(req)
-      || process.env.APOS_ASTRO_STATIC_BUILD === '1'
-      || Boolean(config.staticBuild);
-    const input = (isStaticBuild && req && typeof req !== 'string') ? req.url : req;
-    let request = aposRequest(input);
-    if (request.method === 'HEAD') {
-      request = new Request(request, {
-        method: 'GET'
-      });
-    }
-    const response = await aposResponse(request);
-    let headers = response.headers;
-    aposData = await response.json();
-
-    // Apostrophe's external-front middleware returns redirects as JSON
-    // (e.g. { redirect: true, url: '/fr/', status: 302 }). When the
-    // redirect only adds or removes a trailing slash we should follow
-    // it internally rather than bouncing the browser — otherwise locale
-    // home pages like /fr/ cause an infinite redirect loop.
-    // Skip the site root "/" — it never needs this treatment.
-    //
-    // When a prefix is configured, Apostrophe returns redirect URLs
-    // without the prefix (it's a routing concern, not stored in page
-    // data).  Strip the prefix from `from` so both sides compare on
-    // the same terms, then re-add it when constructing the retry URL.
-    if (aposData.redirect && aposData.url !== '/') {
-      const prefix = config.aposPrefix || '';
-      let from = new URL(request.url).pathname.replace(/\/+$/, '');
-      if (prefix && from.startsWith(prefix + '/')) {
-        from = from.slice(prefix.length);
-      } else if (prefix && from === prefix) {
-        from = '/';
-      }
-      const to = (aposData.url || '').replace(/\/+$/, '');
-      if (from === to) {
-        const retryUrl = prefix + aposData.url;
-        const retry = new Request(new URL(retryUrl, request.url), request);
-        const retryResponse = await aposResponse(retry);
-        headers = retryResponse.headers;
-        const retryData = await retryResponse.json();
-        // Safety check: if the retry itself redirects to the same
-        // URL we just tried, we've hit an infinite redirect loop.
-        // Return an error instead of bouncing forever.
-        if (retryData.redirect && retryData.url === aposData.url) {
-          throw new Error(
-            `Infinite redirect detected: ${aposData.url} redirects back to itself`
-          );
-        }
-        aposData = retryData;
-      }
-    }
-
-    aposData.aposResponseHeaders = headers;
-    if (aposData.template === '@apostrophecms/page:notFound') {
-      aposData.notFound = true;
-    }
-  } catch (e) {
-    console.error('error:', e);
-    aposData.errorFetchingPage = e;
-    aposData.page = {
-      type: 'apos-fetch-error'
-    };
-  }
-  return aposData;
-}
