+++ packages/apostrophe-astro/helpers/server/index.js
@@ -9,6 +9,6 @@
  * @module @apostrophecms/apostrophe-astro/helpers/server
  */
 
-export { aposFetch, aposPageFetch } from './fetch.js';
+export { aposFetch } from './fetch.js';
 export { getAposHost, isStaticBuild } from './url.js';
 export { getAllStaticPaths, getAllUrlMetadata, getLocales } from './static.js';
