+++ packages/apostrophe-astro/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@apostrophecms/apostrophe-astro",
-  "version": "1.13.0",
+  "version": "1.12.0",
   "type": "module",
   "description": "Apostrophe integration for Astro",
   "repository": {
