+++ packages/apostrophe-astro/types/helpers/index.d.ts
@@ -1,2 +1,2 @@
-export { aposFetch, aposPageFetch, getAposHost, isStaticBuild, getAllStaticPaths, getAllUrlMetadata, getLocales } from "./server/index.js";
+export { aposFetch, getAposHost, isStaticBuild, getAllStaticPaths, getAllUrlMetadata, getLocales } from "./server/index.js";
 export { buildPageUrl, getFilterBaseUrl, aposSetQueryParameter, slugify, stylesElements, stylesAttributes, getFocalPoint, getAttachmentUrl, getAttachmentSrcset, getWidth, getHeight } from "./universal/index.js";
