+++ packages/apostrophe-astro/types/helpers/server/index.d.ts
@@ -1,3 +1,3 @@
-export { aposFetch, aposPageFetch } from "./fetch.js";
+export { aposFetch } from "./fetch.js";
 export { getAposHost, isStaticBuild } from "./url.js";
 export { getAllStaticPaths, getAllUrlMetadata, getLocales } from "./static.js";
