+++ packages/apostrophe-astro/types/lib/aposPageFetch.d.ts
@@ -1,1 +1,6 @@
-export { aposPageFetch as default } from "../helpers/server/fetch.js";
+/**
+ * @internal For use in the starter kit's `[...slug].astro` entrypoint only.
+ * Not part of the public helper API.
+ */
+export declare function aposPageFetch(req: Request): Promise<object>;
+export default aposPageFetch;
